// examples/arithmetics-lezer/src/language-server/generated/ast.ts
export type AbstractDefinition = DeclaredParameter | Definition;

export const AbstractDefinition = {
    $type: 'AbstractDefinition'
} as const;

export function isAbstractDefinition(item: unknown): item is AbstractDefinition {
    return reflection.isInstance(item, AbstractDefinition.$type);
}

export interface BinaryExpression extends langium.AstNode {
    readonly $container: BinaryExpression | Definition | Evaluation | FunctionCall;
    readonly $type: 'BinaryExpression';
    left: Expression;
    operator: '%' | '*' | '+' | '-' | '/' | '^';
    right: Expression;
}

export const BinaryExpression = {
    $type: 'BinaryExpression',
    left: 'left',
    operator: 'operator',
    right: 'right'
} as const;
// ...
export interface Definition extends langium.AstNode {
    readonly $container: Module;
    readonly $type: 'Definition';
    args: Array<DeclaredParameter>;
    expr: BinaryExpression;
    name: string;
}

export const Definition = {
    $type: 'Definition',
    args: 'args',
    expr: 'expr',
    name: 'name'
} as const;

export function isDefinition(item: unknown): item is Definition {
    return reflection.isInstance(item, Definition.$type);
}

export interface Evaluation extends langium.AstNode {
    readonly $container: Module;
    readonly $type: 'Evaluation';
    expression: BinaryExpression;
}

export const Evaluation = {
    $type: 'Evaluation',
    expression: 'expression'
} as const;

export function isEvaluation(item: unknown): item is Evaluation {
    return reflection.isInstance(item, Evaluation.$type);
}

export type Expression = BinaryExpression | FunctionCall | NumberLiteral;

export const Expression = {
    $type: 'Expression'
} as const;

export function isExpression(item: unknown): item is Expression {
    return reflection.isInstance(item, Expression.$type);
}

export interface FunctionCall extends langium.AstNode {
    readonly $container: BinaryExpression;
    readonly $type: 'FunctionCall';
    args: Array<BinaryExpression>;
    func: langium.Reference<AbstractDefinition>;
}

export const FunctionCall = {
    $type: 'FunctionCall',
    args: 'args',
    func: 'func'
} as const;

export function isFunctionCall(item: unknown): item is FunctionCall {
    return reflection.isInstance(item, FunctionCall.$type);
}

export interface Module extends langium.AstNode {
    readonly $type: 'Module';
    name: string;
    statements: Array<Statement>;
}

export const Module = {
    $type: 'Module',
    name: 'name',
    statements: 'statements'
} as const;

export function isModule(item: unknown): item is Module {
    return reflection.isInstance(item, Module.$type);
}

export interface NumberLiteral extends langium.AstNode {
    readonly $container: BinaryExpression;
    readonly $type: 'NumberLiteral';
    value: number;
}

export const NumberLiteral = {
    $type: 'NumberLiteral',
    value: 'value'
} as const;

export function isNumberLiteral(item: unknown): item is NumberLiteral {
    return reflection.isInstance(item, NumberLiteral.$type);
}

export type Statement = Definition | Evaluation;

export const Statement = {
    $type: 'Statement'
} as const;